Move rake notes test to railties/test/application/rake/notes_test.rb
